test(scanner): clarify fragment slice coverage
