+++ rust/lance/src/dataset/tests/dataset_overlay_index_masking.rs
@@ -260,7 +260,7 @@ async fn test_overlay_stale_drop_and_new_match(#[values(false, true)] stable_row
 /// constrain it too, otherwise an overlaid row outside the requested slice can be reintroduced.
 #[rstest]
 #[tokio::test]
-async fn test_physical_row_selection_constrains_overlay_stale_replay(
+async fn test_overlay_matches_respect_physical_row_selection(
     #[values(false, true)] stable_row_ids: bool,
 ) {
     let mut dataset = create_base_dataset_with(stable_row_ids).await;
@@ -276,20 +276,20 @@ async fn test_physical_row_selection_constrains_overlay_stale_replay(
     )
     .await;
 
-    let selected_addr = u64::from(RowAddress::new_from_parts(0, 2));
+    let unselected_row_addr = u64::from(RowAddress::new_from_parts(0, 2));
     let mut scanner = dataset.scan();
     scanner
-        .with_physical_row_addr_prefilter(RowAddrTreeMap::from_iter([selected_addr]))
+        .with_physical_row_addr_prefilter(RowAddrTreeMap::from_iter([unselected_row_addr]))
         .filter("age = 999")
         .unwrap()
         .project(&["id"])
         .unwrap();
     assert_eq!(scanner.try_into_batch().await.unwrap().num_rows(), 0);
 
-    let selected_addr = u64::from(RowAddress::new_from_parts(0, 1));
+    let updated_row_addr = u64::from(RowAddress::new_from_parts(0, 1));
     let mut scanner = dataset.scan();
     scanner
-        .with_physical_row_addr_prefilter(RowAddrTreeMap::from_iter([selected_addr]))
+        .with_physical_row_addr_prefilter(RowAddrTreeMap::from_iter([updated_row_addr]))
         .filter("age = 999")
         .unwrap()
         .project(&["id"])
